+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
 ### Fixed
 
 - Fixed the hook scope in the disguiser causing an error
+- Fixed the classic entity spawn mode breaking on maps without all three spawn types
 
 ## [v0.10.2b](https://github.com/TTT-2/TTT2/tree/v0.10.2b) (2021-10-21)